ディレクトリ構成の変更
disk,cell,board,gameのクラスをそれぞれのファイルに分離
