+++ lib/app/othello/widget/board.dart
@@ -1,58 +1,8 @@
 import 'package:flutter/material.dart';
 
-//マス目
-class OthelloSquare extends StatelessWidget {
-  const OthelloSquare({
-    Key? key,
-    required this.onTap,
-    required this.value,
-  }) : super(key: key);
+import 'package:othello/app/othello/widget/cell.dart';
 
-  final void Function() onTap;
-  final String? value;
-
-  @override
-  Widget build(BuildContext context) {
-    //タップの入力検知
-    return GestureDetector(
-      onTap: onTap,
-      child: Container(
-        //マス目の区切り線
-        decoration: BoxDecoration(
-          border: Border.all(color: const Color.fromARGB(255, 255, 255, 255), width: 0.1),
-        ),
-        child: Center(
-            //マス目の中に表示するもの
-            child:
-                //Text(value!),
-                _othelloDiscJudg(context, value)),
-      ),
-    );
-  }
-
-  //石の色分け
-  Widget? _othelloDiscJudg(BuildContext context, var value) {
-    if (value == "B") {
-      return _othelloDisc(context, Colors.black);
-    } else if (value == "W") {
-      return _othelloDisc(context, Colors.white);
-    }
-    return null;
-  }
-
-  //石のUI
-  Widget _othelloDisc(BuildContext context, Color color) {
-    return Container(
-      margin: const EdgeInsets.all(4),
-      decoration: BoxDecoration(
-        color: color,
-        shape: BoxShape.circle,
-      ),
-    );
-  }
-}
-
-//マス目[OthelloSquare]を表示するBoard
+//マス目[OthelloCell]を表示するBoard
 class OthelloBoard extends StatefulWidget {
   const OthelloBoard({
     Key? key,
@@ -135,7 +85,7 @@ class _OthelloBoardState extends State<OthelloBoard> {
               //表示するSquareの数
               children: List.generate(
                   64,
-                  (int i) => OthelloSquare(
+                  (int i) => OthelloCell(
                         onTap: () => handleClick(i),
                         value:
                             //'$i' テキストを表示するだけの値
@@ -155,34 +105,6 @@ class _OthelloBoardState extends State<OthelloBoard> {
               ],
             )),
       ],
-    );
-  }
-}
-
-//盤面[OthelloBoard]を表示するGame
-class OthelloGame extends StatelessWidget {
-  const OthelloGame({Key? key}) : super(key: key);
-
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      home: Scaffold(
-        appBar: AppBar(
-          title: const Text('Otello'),
-        ),
-        body: Center(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: <Widget>[
-              const OthelloBoard(),
-              Padding(
-                padding: const EdgeInsets.only(left: 20),
-                child: Column(children: const []),
-              ),
-            ],
-          ),
-        ),
-      ),
     );
   }
 }
